+++ src/ui/common_ui.c
@@ -124,6 +124,7 @@ void transfer_settings_to_ui(void)
         ui->coverAnsi = (settings->coverAnsi[0] == '1');
         ui->hideHelp = (settings->hideHelp[0] == '1');
         ui->visualizerEnabled = (settings->visualizerEnabled[0] == '1');
+        ui->discordRPCEnabled = (settings->discordRPCEnabled[0] == '1');
         ui->quitAfterStopping = (settings->quitAfterStopping[0] == '1');
         ui->hideGlimmeringText = (settings->hideGlimmeringText[0] == '1');
         ui->mouseEnabled = (settings->mouseEnabled[0] == '1');
@@ -139,7 +140,6 @@ void transfer_settings_to_ui(void)
         ui->allowNotifications = (settings->allowNotifications[0] == '1');
         ui->coverEnabled = (settings->coverEnabled[0] == '1');
         ui->coverAnsi = (settings->coverAnsi[0] == '1');
-        ui->visualizerEnabled = (settings->visualizerEnabled[0] == '1');
         ui->shuffle_enabled = (settings->shuffle_enabled[0] == '1');
 
         int tmp_num_bytes =